cloud/end_to_end_automated_pipeline/components/data_ingestion_container/requirements.txt
google-cloud-storage
pandas
google-cloud-bigquery